+++ cfg/config.php
@@ -2,10 +2,10 @@
 
 define('JOM_DEBUG',     true);
 define('JOM_LOG',       true);
-define('JOM_MILESTONE', '0.1');                         // installer version (refer to MILESTONES.md for versioning progression)
-define('JOM_RELEASE',   '10');
-define('JOM_VERSION',   JOM_MILESTONE.'.'.JOM_RELEASE); // milestone.release
-define('JOM_DESC_VER',  'v. '.JOM_MILESTONE.'.x: installation procedure / database, tables and sample data');   // milestone version description
+define('JOM_MILESTONE', '0.2');                                         // LOGIN version (refer to MILESTONES.md for versioning progression)
+define('JOM_RELEASE',   '2');
+define('JOM_VERSION',   JOM_MILESTONE.'.'.JOM_RELEASE);                 // milestone.release
+define('JOM_DESC_VER',  'v. '.JOM_MILESTONE.'.x: login procedures');    // milestone version description
 
 
 // directories definition
@@ -32,4 +32,5 @@ define('TBL_JOBS',                TABLES_PREFIX.'Jobs');
 define('TBL_USERS_ACL',           TABLES_PREFIX.'Users_acl');
 define('TBL_CATEGORIES_1',        TABLES_PREFIX.'Categories_1');
 define('TBL_CATEGORIES_2',        TABLES_PREFIX.'Categories_2');
-define('TBL_LOGGER',              TABLES_PREFIX.'log');
+define('TBL_LOGIN_ATTEMPTS',      TABLES_PREFIX.'Login_attempts');
+define('TBL_LOGGER',              TABLES_PREFIX.'Log');